\name{CurlFeatureBits}
\alias{CurlFeatureBits}
\alias{CurlNetrc}
\alias{CurlGlobalBits}
\docType{data}
\title{Constants for libcurl}
\description{
  These are enums and bit fields defining
  constants used in libcurl and used in
  R to specify values symbolically.
}
\usage{CurlFeatureBits}
\format{
  named integer vectors.
  The names give the symbolic constants that can be used
  in R and C code. These are mapped to their integer equivalents
  and used in C-level computations.
}
\source{
  libcurl (see \url{http://curl.haxx.se})
}
%\references{}
%\seealso{}
%\examples{}
\keyword{datasets}
