\name{CURLEnums}
\alias{|,BitwiseValue,BitwiseValue-method}
\alias{c,BitwiseValue-method}
\alias{&,BitwiseValue-method}
\alias{&,BitwiseValue,BitwiseValue-method}
\alias{[,EnumDef-method}
\alias{[,EnumDef,ANY-method}
\alias{coerce,BitwiseValue,numeric-method}
\alias{curl_infotype-class}
\alias{CURLcode-class}
\alias{curl_proxytype-class}
\alias{curl_usessl-class}
\alias{curl_ftpccc-class}
\alias{curl_ftpauth-class}
\alias{curl_ftpcreatedir-class}
\alias{curl_ftpmethod-class}
\alias{CURL_NETRC_OPTION-class}
\alias{CURLFORMcode-class}
\alias{curl_TimeCond-class}
\alias{curl_closepolicy-class}
\alias{coerce,integer,curl_infotype-method}
\alias{coerce,integer,CURLcode-method}
\alias{coerce,integer,curl_proxytype-method}
\alias{coerce,integer,curl_usessl-method}
\alias{coerce,integer,curl_ftpccc-method}
\alias{coerce,integer,curl_ftpauth-method}
\alias{coerce,integer,curl_ftpcreatedir-method}
\alias{coerce,integer,curl_ftpmethod-method}
\alias{coerce,integer,CURL_NETRC_OPTION-method}
\alias{coerce,integer,CURLFORMcode-method}
\alias{coerce,integer,curl_TimeCond-method}
\alias{coerce,integer,curl_closepolicy-method}
\alias{coerce,numeric,curl_infotype-method}
\alias{coerce,numeric,CURLcode-method}
\alias{coerce,numeric,curl_proxytype-method}
\alias{coerce,numeric,curl_usessl-method}
\alias{coerce,numeric,curl_ftpccc-method}
\alias{coerce,numeric,curl_ftpauth-method}
\alias{coerce,numeric,curl_ftpcreatedir-method}
\alias{coerce,numeric,curl_ftpmethod-method}
\alias{coerce,numeric,CURL_NETRC_OPTION-method}
\alias{coerce,numeric,CURLFORMcode-method}
\alias{coerce,numeric,curl_TimeCond-method}
\alias{coerce,numeric,curl_closepolicy-method}
\alias{coerce,character,curl_infotype-method}
\alias{coerce,character,CURLcode-method}
\alias{coerce,character,curl_proxytype-method}
\alias{coerce,character,curl_usessl-method}
\alias{coerce,character,curl_ftpccc-method}
\alias{coerce,character,curl_ftpauth-method}
\alias{coerce,character,curl_ftpcreatedir-method}
\alias{coerce,character,curl_ftpmethod-method}
\alias{coerce,character,CURL_NETRC_OPTION-method}
\alias{coerce,character,CURLFORMcode-method}
\alias{coerce,character,curl_TimeCond-method}
\alias{coerce,character,curl_closepolicy-method} 
\title{Classes and coercion methods for enumerations in libcurl}
\description{These are classes and coercion methods
for enumeration types in RCurl corresponding to symbolic
constants in libcurl.
The actual constants are not exported, but are defined within
the package. So we can use them with code such as 
\code{RCurl:::CURLINFO_DATA_IN}.
}
%\usage{}
%\value{}
\author{Duncan Temple Lang}
\keyword{programming}
