\name{HTTP_VERSION_1_0}
\Rdversion{1.1}
\alias{HTTP_VERSION_1_0}
 \alias{HTTP_VERSION_1_1}
 \alias{HTTP_VERSION_LAST}
 \alias{HTTP_VERSION_NONE}
\alias{SSLVERSION_DEFAULT}
 \alias{SSLVERSION_LAST}
 \alias{SSLVERSION_SSLv2}
 \alias{SSLVERSION_SSLv3}
 \alias{SSLVERSION_TLSv1}
\title{Symbolic constants for specifying HTTP and SSL versions in libcurl}
\description{
These are values that can be used to set the \code{http.version} and \code{sslversion}
options of \code{\link{curlPerform}}.
}
\usage{HTTP_VERSION_1_0}
\details{
}
\references{
 \url{http://curl.haxx.se/libcurl/c/curl_easy_setopt.html}
}
\examples{
}
\keyword{datasets}
